+++ src/com/dre/brewery/filedata/BData.java
@@ -2,9 +2,7 @@ package com.dre.brewery.filedata;
 
 import com.dre.brewery.*;
 import com.dre.brewery.lore.Base91DecoderStream;
-import com.dre.brewery.recipe.CustomItem;
 import com.dre.brewery.recipe.Ingredient;
-import com.dre.brewery.recipe.PluginItem;
 import com.dre.brewery.recipe.SimpleItem;
 import com.dre.brewery.utility.BUtil;
 import com.dre.brewery.utility.BoundingBox;
@@ -22,7 +20,6 @@ import java.io.File;
 import java.io.IOException;
 import java.util.*;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.stream.IntStream;
 
 public class BData {
