+++ src/com/dre/brewery/listeners/BlockListener.java
@@ -1,9 +1,6 @@
 package com.dre.brewery.listeners;
 
-import com.dre.brewery.BPlayer;
-import com.dre.brewery.BSealer;
-import com.dre.brewery.Barrel;
-import com.dre.brewery.DistortChat;
+import com.dre.brewery.*;
 import com.dre.brewery.P;
 import com.dre.brewery.api.events.barrel.BarrelDestroyEvent;
 import com.dre.brewery.filedata.BConfig;
