+++ src/com/dre/brewery/listeners/CauldronListener.java
@@ -1,7 +1,6 @@
 package com.dre.brewery.listeners;
 
 import com.dre.brewery.BCauldron;
-import com.dre.brewery.P;
 import com.dre.brewery.utility.LegacyUtil;
 import org.bukkit.Material;
 import org.bukkit.block.BlockState;
